fix math

+++ src/platforms/arm/int.h
@@ -16,12 +16,12 @@ namespace fl {
     // - STM32F1 (Maple Mini and similar)
     // - Arduino GIGA R1 (STM32H747)
     // - Nordic nRF52 family (nRF52832, nRF52840, etc.)
-    typedef short i16;
-    typedef unsigned short u16;
-    typedef long i32;
-    typedef unsigned long u32;
-    typedef long long i64;
-    typedef unsigned long long u64;
+    typedef int16_t i16;
+    typedef uint16_t u16;
+    typedef int32_t i32;
+    typedef uint32_t u32;
+    typedef int64_t i64;
+    typedef uint64_t u64;
     // size_t is unsigned long on ARM (32-bit)
     typedef size_t size;
     // uintptr_t is unsigned long on ARM (32-bit pointers)
